updated-mysql-task/project/requirements.txt
Flask
Flask-RESTful
requests
mysql-connector-python